Starting Review Framework
Modify the cuda path in the Makefile to run the dpct on olice.
Run the dpct, the commands used are:

```
intercept-build make cudatest
dpct -p compile_commands.json -in-root=src/cudatest/ -out-root=src/mysycltest
```

The output of the dpct is stored as is in the src/mysycltest directory.
Now a review of the translated code is needed, maybe with a comparison with
the src/sycltect directory. The question is: is there something to update?